Display: add support for the HD44780 display.
All in one chunk because the infrastructure is already there.
This also implements the parallel 4-bit bus used by quite some
displays.

For now you have to add quite a number of #defines to your
config.h. First, there are all the pins required, pin names
changed to your actual board/display connection, of course:

  #define DISPLAY_BUS_PIN_RS       PC1
  #define DISPLAY_BUS_PIN_RW       PC0
  #define DISPLAY_BUS_PIN_E        PD2
  #define DISPLAY_BUS_PIN_D4       PD3
  #define DISPLAY_BUS_PIN_D5       PD4
  #define DISPLAY_BUS_PIN_D6       PD5
  #define DISPLAY_BUS_PIN_D7       PD6

And then the information about the display actually existing:

  #define DISPLAY_BUS_4BIT
  #define DISPLAY_TYPE_HD44780

Allowing to do all this in Configtool is forthcoming, of course.